+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     - angle between 2 float2 vectors
   - float3 extensions
     - float3 LongAxis (select axis with the largest value)
+  - mathx
+    - approximately using Unity.Mathematics instead of Mathf
 
 ## [1.1.0]